requirements.txt
fastapi
uvicorn
python-multipart
pillow
rembg
onnxruntime
